[FIX] l10n_sa_edi: fix dependencies
l10n_sa_edi uses account_edi methods but has no dependency
on it. It still works since for the moment account_edi is
auto_installed with account but it won't be the case starting
from saas-16.4.

closes odoo/odoo#131730

Related: #130836
X-original-commit: f70b052
