+++ addons/account_edi/models/__init__.py
@@ -6,4 +6,3 @@ from . import account_edi_format
 from . import account_edi_document
 from . import ir_actions_report
 from . import ir_attachment
-from . import uom
