+++ addons/l10n_sa_edi/__manifest__.py
@@ -6,6 +6,7 @@
     'countries': ['sa'],
     'version': '0.1',
     'depends': [
+        'account_edi',
         'account_edi_ubl_cii',
         'account_debit_note',
         'l10n_sa',
